Add new Database category to DefaultCategories
A new static readonly `Category` property named `Database`
was added to the `DefaultCategories` class. This category
has the following properties:
- `Id`: "system.database"
- `ResourceName`: "Alma_Category_Database"
- `DefaultName`: "Banco de Dados"
- `IsSystemDefault`: true

The `Database` category was also added to the `All` list of
categories, ensuring it is included in the collection of all
default categories. This change introduces a new system
default category to better organize and represent database-
related workflows.